+++ gems/quilt_rails/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Added
+
+- The `Quilt::ReactRenderable` method `render_react` now accepts a header argument that gets passed through the reverse proxy
+
 ## [1.9.2] - 2019-12-10
 
 ## Changed